init start to move to integration tests
 On branch feature/init-integration-switch
 Changes to be committed:
	modified:   .circleci/config.yml
	modified:   README.md
	modified:   package.json
	renamed:    src/__tests__/App.test.tsx -> src/__tests__/unit/App.test.tsx
	renamed:    src/__tests__/context.test.tsx -> src/__tests__/unit/context.test.tsx
	renamed:    src/__tests__/pages/Favourite.test.tsx -> src/__tests__/unit/pages/Favourite.test.tsx
	renamed:    src/__tests__/pages/Home.test.tsx -> src/__tests__/unit/pages/Home.test.tsx
	renamed:    src/__tests__/pages/account/Login.test.tsx -> src/__tests__/unit/pages/account/Login.test.tsx
	renamed:    src/__tests__/pages/account/Settings.test.tsx -> src/__tests__/unit/pages/account/Settings.test.tsx
	renamed:    src/__tests__/pages/account/Signup.test.tsx -> src/__tests__/unit/pages/account/Signup.test.tsx
	renamed:    src/__tests__/pages/media/MediaDetails.test.tsx -> src/__tests__/unit/pages/media/MediaDetails.test.tsx
	renamed:    src/__tests__/pages/media/Search.test.tsx -> src/__tests__/unit/pages/media/Search.test.tsx
	modified:   src/setupTests.ts
