+++ src/__tests__/unit/App.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, render, waitForElement } from '@testing-library/react'
 import React from 'react';
-import App from '../App';
+import App from '../../App';
 
 describe('Main App', () => {
   afterEach(cleanup);
