+++ src/__tests__/unit/pages/Favourite.test.tsx
@@ -1,8 +1,8 @@
 import { cleanup, render, waitForElement } from '@testing-library/react'
 import React, { Suspense } from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import UserContext, { init } from '../../context';
-import FavouritePage from '../../pages/Favourite';
+import UserContext, { init } from '../../../context';
+import FavouritePage from '../../../pages/Favourite';
 
 const favs = [
   {
