+++ src/__tests__/unit/pages/account/Login.test.tsx
@@ -2,8 +2,8 @@ import { cleanup, fireEvent, render } from '@testing-library/react';
 import { createMemoryHistory } from 'history';
 import React from 'react';
 import { MemoryRouter, Router } from 'react-router';
-import UserContext, { init } from '../../../context';
-import LoginPage from '../../../pages/account/Login';
+import UserContext, { init } from '../../../../context';
+import LoginPage from '../../../../pages/account/Login';
 
 describe('Login Page', () => {
   afterEach(cleanup);
