+++ src/__tests__/unit/pages/account/Signup.test.tsx
@@ -3,8 +3,8 @@ import * as firebase from 'firebase/app';
 import { createMemoryHistory } from 'history';
 import React from 'react';
 import { MemoryRouter, Router } from 'react-router';
-import UserContext, { init } from '../../../context';
-import Signup from '../../../pages/account/Signup';
+import UserContext, { init } from '../../../../context';
+import Signup from '../../../../pages/account/Signup';
