+++ src/setupTests.ts
@@ -1,3 +1,2 @@
 import { mockIonicReact } from '@ionic/react-test-utils';
-import '@testing-library/jest-dom/extend-expect';
 mockIonicReact();
